removed dom util, moved functions into relevant modules  #495
