+++ src/js/leaderboards.js
@@ -1,5 +1,5 @@
 import * as CloudFunctions from "./cloud-functions";
-import { showBackgroundLoader, hideBackgroundLoader } from "./dom-util";
+import * as Loader from "./loader";
 import * as Notifications from "./notification-center";
 
 let currentLeaderboard = "time_15";
@@ -17,7 +17,6 @@ export function hide() {
         $("#leaderboardsWrapper").addClass("hidden");
       }
     );
-  // focusWords();
 }
 
 function update() {
@@ -33,7 +32,7 @@ function update() {
     uid = firebase.auth().currentUser.uid;
   }
 
-  showBackgroundLoader();
+  Loader.show();
   Promise.all([
     CloudFunctions.getLeaderboard({
       mode: boardinfo[0],
@@ -49,7 +48,7 @@ function update() {
     }),
   ])
     .then((lbdata) => {
-      hideBackgroundLoader();
+      Loader.hide();
       let dailyData = lbdata[0].data;
       let globalData = lbdata[1].data;
 
@@ -245,7 +244,7 @@ function update() {
       }
     })
     .catch((e) => {
-      hideBackgroundLoader();
+      Loader.hide();
       Notifications.add("Something went wrong: " + e.message, -1);
     });
 }
