+++ src/js/loader.js
@@ -1,0 +1,7 @@
+export function show() {
+  $("#backgroundLoader").stop(true, true).fadeIn(125);
+}
+
+export function hide() {
+  $("#backgroundLoader").stop(true, true).fadeOut(125);
+}
