+++ src/js/ui.js
@@ -1,5 +1,89 @@
+import Config from "./config";
+
 export let pageTransition = false;
 
 export function setPageTransition(val) {
   pageTransition = val;
+}
+
+export function updateKeytips() {
+  if (Config.swapEscAndTab) {
+    $(".pageSettings .tip").html(`
+    tip: You can also change all these settings quickly using the
+    command line (
+    <key>tab</key>
+    )`);
+    $("#bottom .keyTips").html(`
+    <key>esc</key> - restart test<br>
+      <key>tab</key> - command line`);
+  } else {
+    $(".pageSettings .tip").html(`
+    tip: You can also change all these settings quickly using the
+    command line (
+    <key>esc</key>
+    )`);
+    $("#bottom .keyTips").html(`
+    <key>tab</key> - restart test<br>
+      <key>esc</key> - command line`);
+  }
+}
+
+export function swapElements(
+  el1,
+  el2,
+  totalDuration,
+  callback = function () {
+    return;
+  }
+) {
+  if (
+    (el1.hasClass("hidden") && !el2.hasClass("hidden")) ||
+    (!el1.hasClass("hidden") && el2.hasClass("hidden"))
+  ) {
+    //one of them is hidden and the other is visible
+    if (el1.hasClass("hidden")) {
+      callback();
+      return false;
+    }
+    $(el1)
+      .removeClass("hidden")
+      .css("opacity", 1)
+      .animate(
+        {
+          opacity: 0,
+        },
+        totalDuration / 2,
+        () => {
+          $(el1).addClass("hidden");
+          $(el2)
+            .removeClass("hidden")
+            .css("opacity", 0)
+            .animate(
+              {
+                opacity: 1,
+              },
+              totalDuration / 2,
+              () => {
+                callback();
+              }
+            );
+        }
+      );
+  } else if (el1.hasClass("hidden") && el2.hasClass("hidden")) {
+    //both are hidden, only fade in the second
+    $(el2)
+      .removeClass("hidden")
+      .css("opacity", 0)
+      .animate(
+        {
+          opacity: 1,
+        },
+        totalDuration,
+        () => {
+          callback();
+        }
+      );
+  } else {
+    callback();
+  }
 }
